+++ packages/tce-boot/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @tailor-cms/tce-boot
 
+## 2.1.0
+
+### Minor Changes
+
+- Introduce general feedback.
+
+### Patch Changes
+
+- Updated dependencies
+  - @tailor-cms/tce-display-runtime@2.1.0
+  - @tailor-cms/tce-server-runtime@2.1.0
+  - @tailor-cms/tce-edit-runtime@2.1.0
+  - @tailor-cms/tce-preview-runtime@2.1.0
+
 ## 2.0.3
 
 ### Patch Changes